Add GSConnect role
Basic role, installed and provides GSConnect. Since there is no
`gnome.yml` right now and we only want to install it to Gnome-based setups,
it's not added to `clients.yml`. There will probably be a structural
change that allows to specify more precisely what role is used where.

The role itself downloads the specified version from GitHub and places
it in the global Gnome extension directory. It has to be enabled by the
user. Recommended way is using the Gnome's "Tweaks".